various changes to migrations

+++ db/migrate/20111201024408_create_shipping_methods.rb
@@ -4,8 +4,8 @@ class CreateShippingMethods < ActiveRecord::Migration
       t.belongs_to :shipping_zone,  null: false
       t.string  :name,              null: false
       t.decimal :lower_price_limit, precision: 8, scale: 2
-      t.integer :upper_price_limit, precision: 8, scale: 2
-      t.integer :shipping_price,    precision: 8, scale: 2
+      t.decimal :upper_price_limit, precision: 8, scale: 2
+      t.decimal :shipping_price,    precision: 8, scale: 2
       t.boolean :active,            default: true
 
       t.timestamps
